Rename package of 'help' and 'exit' to global
