+++ src/test/java/seedu/address/logic/parser/SharkieParserTest.java
@@ -22,8 +22,8 @@ import seedu.address.logic.commands.people.PeopleEditCommand;
 import seedu.address.logic.commands.people.PeopleEditCommand.EditPersonDescriptor;
 import seedu.address.logic.commands.people.PeopleFindCommand;
 import seedu.address.logic.commands.people.PeopleListCommand;
-import seedu.address.logic.commands.sharkie.ExitCommand;
-import seedu.address.logic.commands.sharkie.HelpCommand;
+import seedu.address.logic.commands.global.ExitCommand;
+import seedu.address.logic.commands.global.HelpCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.person.NameContainsKeywordsPredicate;
 import seedu.address.model.person.Person;
